+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Note: Prior to version 0.3.4.0, this library was named
 `small-bytearray-builder` is now just a compatibility shim
 to ease the migration process.
 
+## 0.3.5.0 -- 2020-??-??
+
+* Add `wordLEB128` and `word64LEB128`.
+* Add `integerDec` and `naturalDec`.
+* Add `word48PaddedLowerHex`.
+
 ## 0.3.4.0 -- 2020-02-27
 
 * Rename the library from `small-bytearray-builder` to `bytebuild`, and